requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
pandas>=2.2.0
numpy>=1.24.0
pytz>=2024.1
python-dotenv>=1.0.1
supabase>=2.3.0
bcrypt>=4.1.0
groq>=0.9.0
anthropic>=0.40.0
requests>=2.31.0
python-telegram-bot>=21.0
schedule>=1.2.0
cryptography>=42.0
sqlalchemy>=2.0.0
yfinance>=0.2.40
